services/profiles/requirements.txt
strawberry-graphql
fastapi
uvicorn
asyncpg
sqlalchemy
psycopg2-binary